Forbedrer håndtering av loginstatus for utloggingsvarsel
